refactor: Move implementation-specific error to `sqlite_peg_queue.rs`
